refactor: separete components
